Create `logic.rb` and fix linter errors
